transfering code form auth to main
